move functions in cpu_utils.py to utils.py
