+++ python/sglang/srt/model_executor/model_runner.py
@@ -29,7 +29,6 @@ import torch.distributed as dist
 from sglang.srt.configs.device_config import DeviceConfig
 from sglang.srt.configs.load_config import LoadConfig
 from sglang.srt.configs.model_config import AttentionArch, ModelConfig
-from sglang.srt.cpu_utils import get_cpu_ids_by_node
 from sglang.srt.distributed import (
     get_tp_group,
     get_world_group,
@@ -97,6 +96,7 @@ from sglang.srt.utils import (
     enable_show_time_cost,
     get_available_gpu_memory,
     get_bool_env_var,
+    get_cpu_ids_by_node,
     init_custom_process_group,
     is_cuda,
     is_fa3_default_architecture,
